Fix RHEL 9 bugs

+++ addMicrosoftRepo.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
 function has_command() {
-  command -v $1 >/dev/null
+    command -v $1 >/dev/null
 }
 
 # CustomOutput
 function customOutput {
-  figlet -f banner -kc $1 | lolcat
+    figlet -f banner -kc $1 | lolcat
 }
 
 # When no figlet and lolcat
@@ -22,40 +22,44 @@ b_CRER=" \033[1;31m" # bold error color
 b_CWAR=" \033[1;33m"
 # echo like ...  with  flag type  and display message  colors
 prompt() {
-  case ${1} in
-  "-s" | "--success")
-    echo -e "${b_CGSC}${@/-s/}${CDEF}"
-    ;; # print success message
-  "-e" | "--error")
-    echo -e "${b_CRER}${@/-e/}${CDEF}"
-    ;; # print error message
-  "-w" | "--warning")
-    echo -e "${b_CWAR}${@/-w/}${CDEF}"
-    ;; # print warning message
-  "-i" | "--info")
-    echo -e "${b_CCIN}${@/-i/}${CDEF}"
-    ;; # print info message
-  *)
-    echo -e "$@"
-    ;;
-  esac
+    case ${1} in
+        "-s" | "--success")
+            echo -e "${b_CGSC}${@/-s/}${CDEF}"
+        ;; # print success message
+        "-e" | "--error")
+            echo -e "${b_CRER}${@/-e/}${CDEF}"
+        ;; # print error message
+        "-w" | "--warning")
+            echo -e "${b_CWAR}${@/-w/}${CDEF}"
+        ;; # print warning message
+        "-i" | "--info")
+            echo -e "${b_CCIN}${@/-i/}${CDEF}"
+        ;; # print info message
+        *)
+            echo -e "$@"
+        ;;
+    esac
 }
 
 if has_command figlet && has_command lolcat; then
-  customOutput "Add Microsoft repo"
+    customOutput "Add Microsoft repo"
 fi
 prompt -i "Adding Microsoft repo..."
 prompt -i "Addin GPG key..."
 OSID=$(cat /etc/os-release | egrep ^ID= | tr -d ID=\")
 if [[ "$OSID" == "ubuntu" || "$OSID" == "debain" ]]; then
-  curl -sSL https://packages.microsoft.com/keys/microsoft.asc | sudo apt-key add -
+    curl -sSL https://packages.microsoft.com/keys/microsoft.asc | sudo apt-key add -
 fi
 prompt -s "GPG key added!"
 OSID=$(cat /etc/os-release | egrep ^ID= | tr -d ID=\")
 if [[ "$OSID" == "ubuntu" || "$OSID" == "debain" ]]; then
-  sudo apt-add-repository "https://packages.microsoft.com/$OSID/$(lsb_release -a | egrep ^Release: | tr -d Release:)/prod"
+    sudo apt-add-repository "https://packages.microsoft.com/$OSID/$(lsb_release -a | egrep ^Release: | tr -d Release:)/prod"
 fi
 if [[ "$OSID" == "rhel" || "$OSID" == "rocky" || "$OSID" == "centos" || "$OSID" == "almalinux" ]]; then
-  sudo yum install "https://packages.microsoft.com/config/$OSID/$(rpm -E %rhel)/packages-microsoft-prod.rpm"
+    if [[ "$OSID" == "rhel" && "$(rpm -E %rhel)" == 9 ]]; then
+        sudo dnf install "https://packages.microsoft.com/config/$OSID/9/packages-microsoft-prod.rpm"
+    else
+        sudo yum install "https://packages.microsoft.com/config/$OSID/$(rpm -E %rhel)/packages-microsoft-prod.rpm"
+    fi
 fi
 prompt -s "Microsoft repo added!"
